app/requirements.txt
streamlit
pandas
# Pillow


streamlit
Pillow
PyMuPDF
opencv-python
